refactor: merge the same test files with different content
